finish chrono->time refactor
